requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
sqlalchemy==1.4.22
pydantic==1.8.2
psycopg2-binary==2.9.1  # If you are using PostgreSQL
pyjwt
python-dotenv
alembic